+++ app/src/main/java/com/example/todolist/Domain/TaskRepo.kt
@@ -7,6 +7,6 @@ import kotlinx.coroutines.flow.Flow
 interface TaskRepo {
     fun getTasks(): Flow<List<Task>>;
     suspend fun deleteTask(task: Task);
-    suspend fun getTaskById(id:Int) : Task?;
+    suspend fun getTaskById(id:Long) : Task?;
     suspend fun insert(task: Task);
 }
